Add hidden option for a column
In some cases, we need data for one of our data viz widgets but we don’t want to display the records in the table. With the hidden option, we still have access to the data in the query but we skip rendering it in the <th> & <td>  of the <Table> component.