Modify the message related to unavailable CR3 diagrams/narratives
The intent is to not imply that a CR3 is not available when it just
may not have been processed yet. The message is made more general
to try to cover both when a CR3 is just not processed yet, or when
it was not possible for a diagram / narrative to have been extracted.