Update to VZE dashboard link icons and removal of total crashes card
Based on feedback from Lewis, the cards which link to external BI
dashboards and AGOL maps have had thier icons alinged. For the BI
links, due to there not being a brand icon in the FontAwesome
collection a 24x24 px raster icon is added, and the relevent
react components have been extended to support the use of the icon.

Additionally, again based on Lewis' direction, the card indicating
the total number of crashes in the current year has been removed.
Currently, this is done by commenting out that particular component
and related objects, pending a final review by Lewis later this
afternoon.