Resolve comprehensive cost sum for locations
This commit aims to resolve a complex interplay between the Location component
and its useEffect hooks, the get_location_totals() function in the database,
and the comprehensive cost value assigned to non-CR3 crashes.

The get_location_totals() function was reworked recently in an attempt
to address the fact that locations without any non-CR3 crashes were
being allocated one non-CR3 crashes comprehensive cost to the total comp
cost. At the time of writing, this behavior can be seen in production
here: https://visionzero.austin.gov/editor/#/locations/7D3D1117E9.

The change to the function referenced above introduced a performance
issue for the database, and in hind-sight, should not have been merged
into master. To fix this new problem, I proposed a rewrite of the
function which eliminated table joins by using two simple CTE functions.

This rewrite corrected the misbehavior of always returning at least one
non-CR3 crash's worth of comprehensive cost, however, in doing so, it
exposed an issue with one of the useEffect() hooks in the Location
component.

The first useEffect() hook, which is removed by this commit,
was intended to catch the case when the non-CR3 total comp cost was 0, such as
is the case when the function argument for non-cr3 crash comp cost was
set to 0, and augment the variables to include the correct non-CR3 comp
cost per crash to be the value received from the initial query to the
function. However, in the case when a location does not have any non-CR3
crashes, combined with the rewritten function no longer returning
at least one crash's worth of non-CR3 comprehensive cost, the
useEffect() hook would infinitely loop, yielding a WSOD.

This commit's intended resolution is to entirely remove non-CR3 crash
comp cost argument from the function in lieu of a inline subquery in the
fuction which retreives the correct amount out of the
atd_txdot__est_comp_cost table. This removes the need to requery the
database a second time and eliminates the need for this useEffect() hook,
allowing it to be removed which also solves the infinate loop issue it
exhibits for non-CR3 crash bearing locations.

Because the app no longer needs to know the non-CR3 crash
comp cost value, the portion of the GET_LOCATION query which previously
delivered that value can be removed as well from queries/Locations.js.
Essentially this database specific hard-coding of ID 6 has been pushed
down into the database function and out of the application code.

Useful locations for testing this code:

1BD8B11685: 1 CR3 Crash   and 1 Non-CR3 Crashes
7D3D1117E9: 2 CR3 Crashes and 0 Non-CR3 Crashes
95204CB2C7: 0 CR3 Crashes and 1 Non-CR3 Crash