Remove the built-in requirement for crash locations
Previously, crashes had to be tagged as being in the Austin Full Purpose
Jurisdiction or they had to have a city id of 22, and this was enforced
by the graphql query which pulled crashes from the database. Now, with the
forthcoming PR #1007, this logic is pushed down into the database.

A request has been made to make it so that crashes can be searched for
outside of Austin's city limits, and this is accomplished by removing
the above referenced logic from the Crash Query and making it an
advanced search filter.