Remove code repetition & criteria invert directive addition
I had added some unwieldy repetition of code in gqlAbstract.js which is removed.
This previous construct had allowed for a graphql where clause to be
added when a certain toggle was off instead of on. However, that
behavior change was controled by a choice in how you named the where /
where_false, key value pair which was not ideal.

This adds a flag in the advanced filters definition,
invert_toggle_state, which controls if the graphql clause criteria is
added if the state of the toggle is true or false.