Better handle HASURA_SSL_VERIFY flag
Now, if it's not set, as it is currently not set in production, the
script will do the right thing and verify the SSL certificate presented
by Hasura. Generally, this just makes it harder to leave a development
configuration setting in place by accident and have it get to
production.